+++ docs/src/main/java/org/integratedmodelling/docs/highlighting/CreateFunctionDocumentation.java
@@ -1,8 +1,12 @@
 package org.integratedmodelling.docs.highlighting;
 
 /**
- * Run this within the Maven workflow to build the Asciidocs of each
- * function namespace.
+ * Run this within the Maven workflow to build the Asciidocs of each function
+ * namespace.
+ * 
+ * TODO capture the prototype and any additional documentation, calling with
+ * just the namespace and asking for namespaces to the components returned by
+ * Extensions.
  * 
  * @author ferdinando.villa
  *